requirements.txt
blinker==1.4
docutils==0.15.2
feedgenerator==1.9
invoke==1.3.0
Jinja2==2.11.3
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
pelican==4.2.0

Pygments==2.7.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.13.0
tornado==6.0.3
Unidecode==1.1.1
